+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-This release focused on internal refactoring and improving test coverage.
+This release focused on internal refactoring and improving test coverage. Many discoverd bugs have been fixed.
 
 ### Fixed
 - Potential deadlock in bundles processor.
 - Swallowing of exceptions raised within bundler factory and bundler factory streams.
+- Virtual path rules in a config not being correctly checked for duplicate matchers.
 - Other undocumented assorted edge cases.
 
 ## [3.0.0-rc.1] - 2019-01-09